requirements.txt
Django==4.1.2
django-auditlog==2.3.0
djangorestframework==3.14.0
